refactor: share environment diagnostics collector
The environment collector that backs `marimo env` needs to be reused by
a forthcoming HTTP endpoint that serves diagnostics to the feedback UI.
Keeping it in the CLI module would force the server layer to depend on
CLI code, so this moves the collector into a neutral
`marimo._utils.diagnostics` module. `marimo._cli.envinfo` becomes a thin
re-export so the `marimo env` command and its tests are unaffected.

It also adds a `redact_home` option to `get_system_info` and an
`abbreviate_home` helper. UI-facing diagnostics will be shared publicly
in issue reports, so the install location's home-directory prefix is
abbreviated to `~`; the CLI keeps its current raw location by defaulting
`redact_home` to False.
